プレイヤー周りの機能を追加
・ザコ戦からボス戦への切替時の反転処理
・プレイヤーの通常弾のレベルアップ処理
・プレイヤーの移動スピードのレベルアップ処理
・ダメージ時の点滅処理
・ダメージ時に強化が両方1段階下がる処理
・プレイヤーのやられ処理（アニメ画像待ち）

+++ Assets/Scripts/EnemySetting.cs
@@ -13,7 +13,7 @@ public class EnemyData
 {
     public string Id;
     public float Hp;
-    public float Attack;
+    public int Attack;
     public int Money;
 }
